Changed data types

+++ src/com/company/Main.java
@@ -7,20 +7,20 @@ public class Main {
         Scanner scnr = new Scanner(System.in);
         while (flag){
             System.out.println("Enter first argument ");
-            int first = scnr.nextInt();
+            long first = scnr.nextLong();
             System.out.println("Enter second argument ");
-            int second = scnr.nextInt();
-            int result = power(first, second);
+            long second = scnr.nextLong();
+            long result = power(first, second);
             System.out.println("Result is " + result + ".");
             System.out.println("If you want to exit from program, enter 1.");
-            int check = scnr.nextInt();
+            byte check = scnr.nextByte();
             if (check == 1){
                 flag = false;
             }
         }
         System.out.println("GoodBye!");
     }
-    protected static int power(int a, int b) {
+    protected static long power(long a, long b) {
         int res = 1;
         int i = 0;
         while (i<b){
